adding command to generate Kibana security docs
